tweak(stdlib): explicitly convert int to uint8_t when reading byte in xod/uart/soft-uart

+++ workspace/__lib__/xod/uart/soft-uart/patch.cpp
@@ -34,7 +34,7 @@ public:
         int data = _serial.read();
         if (data == -1)
             return false;
-        *out = data;
+        *out = (uint8_t)data;
         return true;
     }
